Mved ImportProgressBar to widgets
