+++ frontend/src/features/widgets.js
@@ -1,6 +1,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React from "react";
 import { Check, Spinner } from "./icons";
+import { useCourseImportProgress } from "../common/api";
 
 export const cssInfoBox =
   "bg-blue-100 border-l-4 border-blue-500 text-blue-700 p-4 mt-6";
@@ -84,3 +85,38 @@ export const LoadingPage = ({ className, children, ...props }) => (
     {children}
   </div>
 );
+
+export function ImportQueueProgressBar({ courseId, defaultTotal, onDone }) {
+  const [cancel, setCancel] = React.useState(false);
+
+  // Ping backend to get status of current import
+  const { data } = useCourseImportProgress(courseId, {
+    onSuccess: ({ working }) => {
+      // We are done, inform the parent
+      if (working.progress >= working.total) {
+        setCancel(true);
+        onDone();
+      }
+    },
+    cancel,
+  });
+  const { working } = data || {};
+  const { progress = 0, total = defaultTotal } = working || {};
+
+  const perc = Math.round((progress / total) * 100);
+  return (
+    <div className="mt-8 mb-8">
+      <div className="relative pt-1 mb-1">
+        <div className="overflow-hidden h-4 text-xs flex rounded bg-blue-200">
+          <div
+            style={{ width: `${perc}%`, transition: "width 3s" }}
+            className="shadow-none flex flex-col text-center whitespace-nowrap text-white justify-center bg-blue-500"
+          />
+        </div>
+      </div>
+      <div className="flex flex-col items-center">
+        <span>{`${progress} of ${total}`}</span>
+      </div>
+    </div>
+  );
+}
